requirements.txt
confusable-homoglyphs==2.0.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0
django-bootstrap3==9.1.0
django-decouple==2.1
django-registration==2.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
static3==0.7.0
whitenoise==3.3.1
